Server/requirements.txt
pycolmap
Flask
Pillow